+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 
 ## [Unreleased]
 
+## [0.5.0] - 2022-10-12
+
 ### Added
 
 - 添加了使用本地模板的功能
@@ -157,7 +159,8 @@
 [语义化版本号]: https://semver.org/spec/v2.0.0.html
 
 <!-- Versions -->
-[unreleased]: https://github.com/batu1579/hamibot-assistant/compare/v0.4.3...HEAD
+[unreleased]: https://github.com/batu1579/hamibot-assistant/compare/v0.5.0...HEAD
+[0.5.0]: https://github.com/batu1579/hamibot-assistant/compare/v0.4.3..v0.5.0
 [0.4.3]: https://github.com/batu1579/hamibot-assistant/compare/v0.4.2..v0.4.3
 [0.4.2]: https://github.com/batu1579/hamibot-assistant/compare/v0.4.1..v0.4.2
 [0.4.1]: https://github.com/batu1579/hamibot-assistant/compare/v0.4.0..v0.4.1